[SYCL][Headers cleanup] Introduce <sycl/detail/is_device_copyable.hpp>
This allows us to drop several `<sycl/buffer.hpp>`'s dependencies.
