+++ sycl/test/include_deps/sycl_buffer.hpp.cpp
@@ -103,16 +103,10 @@
 // CHECK-NEXT: detail/property_helper.hpp
 // CHECK-NEXT: detail/property_list_base.hpp
 // CHECK-NEXT: properties/property_traits.hpp
+// CHECK-NEXT: detail/is_device_copyable.hpp
 // CHECK-NEXT: detail/stl_type_traits.hpp
 // CHECK-NEXT: detail/sycl_mem_obj_allocator.hpp
 // CHECK-NEXT: detail/aligned_allocator.hpp
 // CHECK-NEXT: event.hpp
 // CHECK-NEXT: ext/oneapi/accessor_property_list.hpp
-// CHECK-NEXT: types.hpp
-// CHECK-NEXT: detail/generic_type_traits.hpp
-// CHECK-NEXT: multi_ptr.hpp
-// CHECK-NEXT: detail/memcpy.hpp
-// CHECK-NEXT: detail/vector_convert.hpp
-// CHECK-NEXT: marray.hpp
-// CHECK-NEXT: swizzles.def
 // CHECK-EMPTY:
