+++ sycl/test/include_deps/sycl_detail_core.hpp.cpp
@@ -107,23 +107,24 @@
 // CHECK-NEXT: detail/property_helper.hpp
 // CHECK-NEXT: detail/property_list_base.hpp
 // CHECK-NEXT: properties/property_traits.hpp
+// CHECK-NEXT: detail/is_device_copyable.hpp
 // CHECK-NEXT: detail/stl_type_traits.hpp
 // CHECK-NEXT: detail/sycl_mem_obj_allocator.hpp
 // CHECK-NEXT: detail/aligned_allocator.hpp
 // CHECK-NEXT: event.hpp
 // CHECK-NEXT: ext/oneapi/accessor_property_list.hpp
-// CHECK-NEXT: types.hpp
-// CHECK-NEXT: detail/generic_type_traits.hpp
-// CHECK-NEXT: detail/memcpy.hpp
-// CHECK-NEXT: detail/vector_convert.hpp
-// CHECK-NEXT: marray.hpp
-// CHECK-NEXT: swizzles.def
 // CHECK-NEXT: detail/accessor_iterator.hpp
+// CHECK-NEXT: detail/generic_type_traits.hpp
 // CHECK-NEXT: detail/handler_proxy.hpp
 // CHECK-NEXT: detail/image_accessor_util.hpp
 // CHECK-NEXT: image.hpp
 // CHECK-NEXT: detail/backend_traits.hpp
 // CHECK-NEXT: sampler.hpp
+// CHECK-NEXT: types.hpp
+// CHECK-NEXT: detail/memcpy.hpp
+// CHECK-NEXT: detail/vector_convert.hpp
+// CHECK-NEXT: marray.hpp
+// CHECK-NEXT: swizzles.def
 // CHECK-NEXT: pointers.hpp
 // CHECK-NEXT: properties/accessor_properties.hpp
 // CHECK-NEXT: properties/buffer_properties.hpp
